// frontend/src/api/generated/models/tenantAdminResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */
import type { DeletionEligibility } from './deletionEligibility.ts';

export interface TenantAdminResponse {
  id: number;
  name: string;
  slug: string;
  code: string;
  address: string;
  phone: string;
  is_active: boolean;
  user_count: number;
  deletion_eligibility: DeletionEligibility;
}